requirements.txt
tweepy==3.10.0
selenium==3.141.0
pandas==1.2.4
python-dotenv==0.17.1
